+++ src/main/java/com/simsilica/lemur/component/TextEntryComponent.java
@@ -130,6 +130,7 @@ public class TextEntryComponent extends AbstractGuiComponent
     private boolean singleLine;
     private boolean focused;
     private boolean cursorVisible = true;
+    private Float preferredCursorWidth = null;
 
     private VersionedReference<DocumentModel> modelRef;
     private VersionedReference<Integer> caratRef;
@@ -157,7 +158,7 @@ public class TextEntryComponent extends AbstractGuiComponent
         this.modelRef = model.createReference();
         this.caratRef = model.createCaratReference();
 
-        cursorQuad = new Quad(bitmapText.getLineHeight()/16f, bitmapText.getLineHeight());
+        cursorQuad = new Quad(getCursorWidth(), bitmapText.getLineHeight());
         cursor = new Geometry( "cursor", cursorQuad );
         GuiMaterial mat = GuiGlobals.getInstance().createMaterial(new ColorRGBA(1,1,1,0.75f), false);
         cursor.setMaterial(mat.getMaterial());
@@ -180,7 +181,7 @@ public class TextEntryComponent extends AbstractGuiComponent
         result.preferredSize = null;
         result.textBox = null;
         result.keyHandler = result.new KeyHandler();
-        result.cursorQuad = new Quad(bitmapText.getLineHeight()/16f, bitmapText.getLineHeight());
+        result.cursorQuad = new Quad(getCursorWidth(), bitmapText.getLineHeight());
         result.cursor = new Geometry("cursor", cursorQuad);
         GuiMaterial mat = GuiGlobals.getInstance().createMaterial(new ColorRGBA(1,1,1,0.75f), false);
         result.cursor.setMaterial(mat.getMaterial());
@@ -408,9 +409,33 @@ public class TextEntryComponent extends AbstractGuiComponent
         x *= scale;
         return x;
     }
+    
+    public void setPreferredCursorWidth( Float f ) {
+        this.preferredCursorWidth = f;
+        resizeCursor();
+        resetCursorPosition();
+    }
+    
+    public Float getPreferredCursorWidth() {
+        return preferredCursorWidth;
+    }  
+
+    public float getCursorWidth() {
+        if( preferredCursorWidth != null ) {
+            return preferredCursorWidth;
+        }
+        // Because small cursor widths sometimes make the cursor invisible
+        // for some reason, we'll try to detect the cases where pixels = units
+        // and make sure that the width is never smaller than 1 pixel.
+        float height = bitmapText.getLineHeight();
+        if( height > 5 ) {
+            return Math.max(1, height/16f);
+        } 
+        return height/16f;
+    }
 
     protected void resizeCursor() {
-        cursorQuad.updateGeometry(bitmapText.getLineHeight()/16f, bitmapText.getLineHeight());
+        cursorQuad.updateGeometry(getCursorWidth(), bitmapText.getLineHeight());
         cursorQuad.clearCollisionData(); 
     }
 
@@ -467,7 +492,7 @@ public class TextEntryComponent extends AbstractGuiComponent
             resetCursorState();
         }
 
-        cursor.setLocalTranslation(x, y, 0.01f);
+        cursor.setLocalTranslation(x - getCursorWidth() * 0.5f, y, 0.01f);
     }
 
     public void setText( String text ) {